// src/infra/midtrans/notification/model/indomaret.ts
export type INotifIndomaret = {
  transaction_time: string;
  transaction_status: "capture"|"settlement"|"pending"|"deny"|"cancel"|"expire"|"failure"|"refund"|"partial_refund"|"authorize";
  transaction_id: string;
  store: string;
  status_message: string;
  status_code: string;
  signature_key: string;
  settlement_time: string;
  payment_type: string;
  payment_code: string;
  order_id: string;
  merchant_id: string;
  gross_amount: string;
  currency: string;
  approval_code: string;
};